cmd/sym_dump: add overlay 0 for Biohazard 2 (beta 2)
The size of the overlay is set to the size in bytes of PSX.EXE, and the
address is set to 0x0 for now. Both of these values are guesses, and
those with more insight into the Bio2 project are encouraged to double
check these values.

Note, we add overlay 0 explicitly as the MAIN.SYM file does not contain
an Overlay symbol to describe overlay ID 0.

Updates #2.